feat: generalize the configuration for job metrics endpoint
