+++ zeebe/gateway-protocol/src/main/proto/v2/rest-api.yaml
@@ -366,7 +366,7 @@ paths:
   /system/usage-metrics:
     $ref: 'system.yaml#/paths/~1system~1usage-metrics'
   /system/configuration:
-    $ref: 'system-configuration.yaml#/paths/~1system~1configuration'
+    $ref: 'system.yaml#/paths/~1system~1configuration'
 
   # Tenant endpoints
   /tenants:
