--- zeebe/gateway-protocol/src/main/proto/v2/system-configuration.yaml
@@ -1,91 +1,0 @@
-## Endpoint definitions
-
-paths:
-  /system/configuration:
-    get:
-      tags:
-        - System
-      operationId: getSystemConfiguration
-      x-eventually-consistent: false
-      summary: System configuration
-      description: |
-        Returns the current system configuration. The response is an envelope
-        that groups settings by feature area.
-      responses:
-        "200":
-          description: Current system configuration grouped by feature area.
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/SystemConfigurationResponse'
-              examples:
-                example-1:
-                  value:
-                    jobMetrics:
-                      enabled: true
-                      exportInterval: "PT5M"
-                      maxWorkerNameLength: 100
-                      maxJobTypeLength: 100
-                      maxTenantIdLength: 30
-                      maxUniqueKeys: 9500
-        "401":
-          $ref: 'common-responses.yaml#/components/responses/Unauthorized'
-        "500":
-          $ref: 'common-responses.yaml#/components/responses/InternalServerError'
-
-## Schema definitions
-
-components:
-  schemas:
-
-    SystemConfigurationResponse:
-      description: |
-        Envelope for all system configuration sections. Each property
-        represents a feature area.
-      type: object
-      required:
-        - jobMetrics
-      properties:
-        jobMetrics:
-          $ref: '#/components/schemas/JobMetricsConfigurationResponse'
-
-    JobMetricsConfigurationResponse:
-      description: Configuration for job metrics collection and export.
-      type: object
-      required:
-        - enabled
-        - exportInterval
-        - maxWorkerNameLength
-        - maxJobTypeLength
-        - maxTenantIdLength
-        - maxUniqueKeys
-      properties:
-        enabled:
-          description: Whether job metrics export is enabled.
-          type: boolean
-          example: true
-        exportInterval:
-          description: The interval at which job metrics are exported, as an ISO 8601 duration.
-          type: string
-          example: "PT5M"
-        maxWorkerNameLength:
-          description: The maximum length of the worker name used in job metrics labels.
-          type: integer
-          format: int32
-          example: 100
-        maxJobTypeLength:
-          description: The maximum length of the job type used in job metrics labels.
-          type: integer
-          format: int32
-          example: 100
-        maxTenantIdLength:
-          description: The maximum length of the tenant ID used in job metrics labels.
-          type: integer
-          format: int32
-          example: 30
-        maxUniqueKeys:
-          description: The maximum number of unique metric keys tracked for job metrics.
-          type: integer
-          format: int32
-          example: 9500
-
