--- zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/controller/system/SystemConfigurationController.java
@@ -1,43 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.zeebe.gateway.rest.controller.system;
-
-import io.camunda.gateway.protocol.model.JobMetricsConfigurationResponse;
-import io.camunda.gateway.protocol.model.SystemConfigurationResponse;
-import io.camunda.zeebe.gateway.rest.annotation.CamundaGetMapping;
-import io.camunda.zeebe.gateway.rest.config.GatewayRestConfiguration;
-import io.camunda.zeebe.gateway.rest.config.GatewayRestConfiguration.JobMetricsConfiguration;
-import io.camunda.zeebe.gateway.rest.controller.CamundaRestController;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@CamundaRestController
-@RequestMapping("/v2/system/configuration")
-public class SystemConfigurationController {
-
-  private final GatewayRestConfiguration gatewayRestConfiguration;
-
-  public SystemConfigurationController(final GatewayRestConfiguration gatewayRestConfiguration) {
-    this.gatewayRestConfiguration = gatewayRestConfiguration;
-  }
-
-  @CamundaGetMapping
-  public ResponseEntity<SystemConfigurationResponse> getSystemConfiguration() {
-    final JobMetricsConfiguration jobMetricsCfg = gatewayRestConfiguration.getJobMetrics();
-    final var jobMetricsResponse =
-        new JobMetricsConfigurationResponse()
-            .enabled(jobMetricsCfg.isEnabled())
-            .exportInterval(jobMetricsCfg.getExportInterval().toString())
-            .maxWorkerNameLength(jobMetricsCfg.getMaxWorkerNameLength())
-            .maxJobTypeLength(jobMetricsCfg.getMaxJobTypeLength())
-            .maxTenantIdLength(jobMetricsCfg.getMaxTenantIdLength())
-            .maxUniqueKeys(jobMetricsCfg.getMaxUniqueKeys());
-    final var response = new SystemConfigurationResponse().jobMetrics(jobMetricsResponse);
-    return ResponseEntity.ok(response);
-  }
-}
