+++ zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/controller/system/SystemController.java
@@ -11,12 +11,16 @@ import static io.camunda.zeebe.gateway.rest.mapper.RestErrorMapper.mapErrorToRes
 
 import io.camunda.gateway.mapping.http.search.SearchQueryRequestMapper;
 import io.camunda.gateway.mapping.http.search.SearchQueryResponseMapper;
+import io.camunda.gateway.protocol.model.JobMetricsConfigurationResponse;
+import io.camunda.gateway.protocol.model.SystemConfigurationResponse;
 import io.camunda.gateway.protocol.model.UsageMetricsResponse;
 import io.camunda.search.query.UsageMetricsQuery;
 import io.camunda.security.auth.CamundaAuthenticationProvider;
 import io.camunda.service.UsageMetricsServices;
 import io.camunda.zeebe.gateway.rest.annotation.CamundaGetMapping;
 import io.camunda.zeebe.gateway.rest.annotation.RequiresSecondaryStorage;
+import io.camunda.zeebe.gateway.rest.config.GatewayRestConfiguration;
+import io.camunda.zeebe.gateway.rest.config.GatewayRestConfiguration.JobMetricsConfiguration;
 import io.camunda.zeebe.gateway.rest.controller.CamundaRestController;
 import io.camunda.zeebe.gateway.rest.mapper.RestErrorMapper;
 import org.springframework.http.ResponseEntity;
@@ -25,20 +29,23 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 @CamundaRestController
 @RequiresSecondaryStorage
-@RequestMapping("/v2/system/usage-metrics")
-public class UsageMetricsController {
+@RequestMapping("/v2/system")
+public class SystemController {
 
   private final UsageMetricsServices usageMetricsServices;
   private final CamundaAuthenticationProvider authenticationProvider;
+  private final GatewayRestConfiguration gatewayRestConfiguration;
 
-  public UsageMetricsController(
+  public SystemController(
       final UsageMetricsServices usageMetricsServices,
-      final CamundaAuthenticationProvider authenticationProvider) {
+      final CamundaAuthenticationProvider authenticationProvider,
+      final GatewayRestConfiguration gatewayRestConfiguration) {
     this.usageMetricsServices = usageMetricsServices;
     this.authenticationProvider = authenticationProvider;
+    this.gatewayRestConfiguration = gatewayRestConfiguration;
   }
 
-  @CamundaGetMapping
+  @CamundaGetMapping(path = "/usage-metrics")
   public ResponseEntity<UsageMetricsResponse> getUsageMetrics(
       @RequestParam final String startTime,
       @RequestParam final String endTime,
@@ -47,6 +54,21 @@ public class UsageMetricsController {
 
     return SearchQueryRequestMapper.toUsageMetricsQuery(startTime, endTime, tenantId, withTenants)
         .fold(RestErrorMapper::mapProblemToResponse, this::getMetrics);
+  }
+
+  @CamundaGetMapping(path = "/configuration")
+  public ResponseEntity<SystemConfigurationResponse> getSystemConfiguration() {
+    final JobMetricsConfiguration jobMetricsCfg = gatewayRestConfiguration.getJobMetrics();
+    final var jobMetricsResponse =
+        new JobMetricsConfigurationResponse()
+            .enabled(jobMetricsCfg.isEnabled())
+            .exportInterval(jobMetricsCfg.getExportInterval().toString())
+            .maxWorkerNameLength(jobMetricsCfg.getMaxWorkerNameLength())
+            .maxJobTypeLength(jobMetricsCfg.getMaxJobTypeLength())
+            .maxTenantIdLength(jobMetricsCfg.getMaxTenantIdLength())
+            .maxUniqueKeys(jobMetricsCfg.getMaxUniqueKeys());
+    final var response = new SystemConfigurationResponse().jobMetrics(jobMetricsResponse);
+    return ResponseEntity.ok(response);
   }
 
   private ResponseEntity<UsageMetricsResponse> getMetrics(final UsageMetricsQuery query) {
